Added synchronization of vehicle_look_left and vehicle_look_right controls for vehicle puresync packets when the player is driving an aircraft. Cleaned up server-side CVehicleNames::GetVehicleTypeName function. Incremented netcode version.

+++ trunk/MTA10_Server/mods/deathmatch/logic/CResource.cpp
@@ -2438,7 +2438,7 @@ ResponseCode CResource::HandleRequestCall ( HttpRequest * ipoHttpRequest, HttpRe
                 {
                     std::string strValue ( pEqual + 1, pAnd - (pEqual + 1) );
                     Unescape ( strValue );
-                    if ( iKey + 1 > arguments.size () )
+                    if ( iKey + 1 > static_cast < int > ( arguments.size () ) )
                         arguments.resize ( iKey + 1 );
                     arguments [ iKey ] = strValue;
                 }
